feat: initial field reimplementation
